+++ movie-app/api/core/v1/ratings/models.py
@@ -1,7 +1,9 @@
 from datetime import datetime
+from decimal import Decimal
+
 from sqlmodel import SQLModel
 
 
 class RatingBase(SQLModel):
-    rating: int
+    rating: Decimal
     updated_at: datetime
